Added code to have adjustable time based on number input; start game now starts a new game if a game is not in progress; end game tells who won and lost
Also added side coloration.
Many of the fundamental needs of the game have been addressed here.
I tried to make a method that is similar to the startGame but doesn't use the CountDownTimer. The CountDownTimer is what is causing some of the lag. The score:time ratio is really inconsistent. We will need to get my method working if we want to eliminate lag.