Made better startgame thread; changed app name and icon
The problem with the previous startgame was that the timer was really inconsistent due to the use of the CountDownTimer. The new startgame (with thread) does not use CountDownTimer but instead uses a runnable, which allows it to be more consistent with its clock management.

For the app name and icon changes, I feel like our new logo and name (Real-Time Roshambo) better reflect our values as a company and what our product does.