Added Menu Screen, Adjusted Game Layout, and made tweaks to app behavior
Clicking the play game button on the menu screen will open the rest of the game, where the user can choose to either start the game or quit back to menu. The buttons now become unclickable when the game ends. Player 2 has a timer that is facing them so they can read the time.  A few graphics were added including a logo for the game which now appears on the menu screen (none of the graphics are final).