Added game options screen and made tweaks to code
Added a screen accesible from both the game and the menu where the player can set game time and toggle powerups on or off(not implemented). The options screen can link back to either the menu or the game. ALso the countdownTimerPausable class was deleted as it was no longer in use. Some graphical capabilities have also been added back into the game as it was determined that they were not the source of the lag.