Updates to aesthetics and project structure
Renamed the company domain from whatever random crap it was before to kotassium.com. Changed the minimum API level to 21 so we could do cool stuff with background tints. Added background tints when buttons are unclickable.  Added some more TODOs with major bugs found: sometimes players gain points when they shouldn't be. Also points need to be reset to 0 when a new game is started.