CPU updates
made it so the cpu can only move once per period of time.
this period of time can be changed in the options window, with easy allowing it to move once every 3 seconds, medium 2 seconds, and hard 1 second.

created a method on options to store the previous settings and load them back in after exiting the options page

deleted stuff on menu screen and options screen that was not in use

added a 3.5 second countdown between pressing start game and the game starting

changed all hardcoded strings to android string xml things

Fixed all yellows